+++ UnBBayes/src/unbbayes/datamining/classifiers/decisiontree/C45.java
@@ -397,7 +397,7 @@ public class C45 extends DecisionTreeLearning implements Serializable
 	 * @param instance the instance to be classified
 	 * @return the classification
 	 */
-	public byte classifyInstance(Instance instance)
+	public short classifyInstance(Instance instance)
 	{
           Leaf leaf;
           Node node;
