+++ UnBBayes/src/unbbayes/datamining/gui/neuralnetwork/InferencePanel.java
@@ -218,7 +218,7 @@ public class InferencePanel extends JPanel implements IInferencePanel{
     for (int i = 0; i < numAtt; i++) {
       if (i != classIndex && !instance.isMissing(i)) {
         Attribute att = attArray[i];
-        byte value = instance.getValue(i);
+        short value = instance.getValue(i);
         rule = rule + att.getAttributeName() + ": " + att.value(value) + "\n";
       }
     }
